+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # AllRecipes Crawler Changelog
 
+## v1.1
+- Improved anti-detection code
+- Fixed "no-image" not displaying for missing image recipes
+- Fixed skipping existing recipes should now skip properly
+- Improved filename cleanup / character removal
+
 ## v1.0
 - Initial Release